Refactor devices package and tests
