+++ pyproject.toml
@@ -114,7 +114,7 @@ addopts = [
     "--strict"
 ]
 markers = ["slow: marks tests as slow (deselect with '-m \"not slow\"')"]
-filterwarnings = ["error", "ignore::DeprecationWarning"]
+filterwarnings = ["error", "ignore::DeprecationWarning", "ignore::cryptography.utils.CryptographyDeprecationWarning"]
 # Pytest plugins:
 env = [
     # Set application runtime environment for testing
